rename shape.rs to ir.rs
