+++ lorax/src/lib.rs
@@ -1,10 +1,10 @@
+mod ir;
 mod link;
 mod pool;
 mod rewrite;
-mod shape;
 
-pub use pool::{Pool, Ptr};
-pub use rewrite::{RewriteRule, RewriteRuleSet};
-pub use shape::{
+pub use ir::{
     Block, Constant, OpBuilder, OpResult, Operation, RewritingCtx, Value, Var, rewrite_block,
 };
+pub use pool::{Pool, Ptr};
+pub use rewrite::{RewriteRule, RewriteRuleSet};
